Move top-menu and aside-menu to seperate components. Fixed #58
