+++ src/Nucleus.Web.Vue/src/admin/admin-layout.vue
@@ -1,42 +1,9 @@
 ﻿<template>
     <v-app>
-        <v-navigation-drawer v-model="drawer" fixed clipped app dark>
-            <v-list dense>
-                <v-list-tile v-for="item in mainMenuItems" :key="item.link" :to="item.link">
-                    <v-list-tile-action>
-                        <v-icon>{{ item.icon }}</v-icon>
-                    </v-list-tile-action>
-                    <v-list-tile-content>
-                        <v-list-tile-title>
-                            {{ item.text }}
-                        </v-list-tile-title>
-                    </v-list-tile-content>
-                </v-list-tile>
-                <v-subheader class="mt-3 grey--text text--darken-1">ADMINISTRATION</v-subheader>
-                <v-list>
-                    <v-list-tile v-for="item in adminMenuItems" :key="item.link" :to="item.link">
-                        <v-list-tile-action>
-                            <v-icon>{{ item.icon }}</v-icon>
-                        </v-list-tile-action>
-                        <v-list-tile-content>
-                            <v-list-tile-title>
-                                {{ item.text }}
-                            </v-list-tile-title>
-                        </v-list-tile-content>
-                    </v-list-tile>
-                </v-list>
-            </v-list>
-        </v-navigation-drawer>
-        <v-toolbar color="blue darken-3" dense fixed clipped-left app dark>
-            <v-toolbar-side-icon @click.stop="drawer = !drawer"></v-toolbar-side-icon>
-            <v-toolbar-title>
-                <span class="title">Nucleus Vue</span>
-            </v-toolbar-title>
-            <v-spacer></v-spacer>
-            <v-btn icon v-on:click.native="logOut">
-                <v-icon>exit_to_app</v-icon>
-            </v-btn>
-        </v-toolbar>
+        <aside-menu>
+        </aside-menu>
+        <top-menu>
+        </top-menu>
         <v-content>
             <v-container>
                 <router-view></router-view>
